requirements.txt
tensorflow-gpu
tqdm
bert-for-tf2
sentencepiece
pandas
sklearn
numpy
pyYaml
